Move all classes from `.clear` package to the root
