+++ dtglib/src/main/java/com/kaltura/dtg/ContentManagerImp.java
@@ -1,15 +1,9 @@
-package com.kaltura.dtg.clear;
+package com.kaltura.dtg;
 
 import android.content.Context;
 import android.net.Uri;
 import android.text.TextUtils;
 import android.util.Log;
-
-import com.kaltura.dtg.AppBuildConfig;
-import com.kaltura.dtg.ContentManager;
-import com.kaltura.dtg.DownloadItem;
-import com.kaltura.dtg.DownloadState;
-import com.kaltura.dtg.DownloadStateListener;
 
 import java.io.File;
 import java.util.ArrayList;
