+++ dtglib/src/main/java/com/kaltura/dtg/DefaultDownloadService.java
@@ -1,4 +1,4 @@
-package com.kaltura.dtg.clear;
+package com.kaltura.dtg;
 
 import android.app.Service;
 import android.content.Context;
@@ -14,11 +14,6 @@ import android.text.TextUtils;
 import android.util.Log;
 
 import com.kaltura.android.exoplayer.hls.Variant;
-import com.kaltura.dtg.ContentManager;
-import com.kaltura.dtg.DownloadItem;
-import com.kaltura.dtg.DownloadState;
-import com.kaltura.dtg.DownloadStateListener;
-import com.kaltura.dtg.Utils;
 
 import java.io.File;
 import java.io.IOException;
