+++ dtglib/src/main/java/com/kaltura/dtg/DownloadRequestParams.java
@@ -1,4 +1,4 @@
-package com.kaltura.dtg.clear;
+package com.kaltura.dtg;
 
 import android.net.Uri;
